services/electricity/requirements.txt
fastapi>=0.115.8
uvicorn>=0.34.0
httpx>=0.28.1
asyncpg>=0.30.0
python-dateutil>=2.9.0.post0